+++ chatlas/_github.py
@@ -126,7 +126,7 @@ def ChatGithub(
     ```
     """
     if model is None:
-        model = log_model_default("gpt-4o")
+        model = log_model_default("gpt-4.1")
     if api_key is None:
         api_key = os.getenv("GITHUB_PAT")