+++ chatlas/_tools.py
@@ -2,11 +2,17 @@ from __future__ import annotations
 
 import inspect
 import warnings
-from typing import TYPE_CHECKING, Any, Awaitable, Callable, Optional
+from typing import TYPE_CHECKING, Any, AsyncGenerator, Awaitable, Callable, Optional
 
+import openai
 from pydantic import BaseModel, Field, create_model
 
 from . import _utils
+from ._content import (
+    ContentToolResult,
+    ContentToolResultImage,
+    ContentToolResultResource,
+)
 
 __all__ = (
     "Tool",
@@ -14,6 +20,8 @@ __all__ = (
 )
 
 if TYPE_CHECKING:
+    from mcp import ClientSession as MCPClientSession
+    from mcp import Tool as MCPTool
     from openai.types.chat import ChatCompletionToolParam
 
 
@@ -28,26 +36,168 @@ class Tool:
     ----------
     func
         The function to be invoked when the tool is called.
-    model
-        A Pydantic model that describes the input parameters for the function.
-        If not provided, the model will be inferred from the function's type hints.
-        The primary reason why you might want to provide a model in
-        Note that the name and docstring of the model takes precedence over the
-        name and docstring of the function.
+    name
+        The name of the tool.
+    description
+        A description of what the tool does.
+    parameters
+        A dictionary describing the input parameters and their types.
     """
 
     func: Callable[..., Any] | Callable[..., Awaitable[Any]]
 
     def __init__(
         self,
+        *,
+        func: Callable[..., Any] | Callable[..., Awaitable[Any]],
+        name: str,
+        description: str,
+        parameters: dict[str, Any],
+    ):
+        self.name = name
+        self.func = func
+        self._is_async = _utils.is_async_callable(func)
+        self.schema: "ChatCompletionToolParam" = {
+            "type": "function",
+            "function": {
+                "name": name,
+                "description": description,
+                "parameters": parameters,
+            },
+        }
+
+    @classmethod
+    def from_func(
+        cls: type["Tool"],
         func: Callable[..., Any] | Callable[..., Awaitable[Any]],
         *,
         model: Optional[type[BaseModel]] = None,
-    ):
-        self.func = func
-        self._is_async = _utils.is_async_callable(func)
-        self.schema = func_to_schema(func, model)
-        self.name = self.schema["function"]["name"]
+    ) -> "Tool":
+        """
+        Create a Tool from a Python function
+
+        Parameters
+        ----------
+        func
+            The function to wrap as a tool.
+        model
+            A Pydantic model that describes the input parameters for the function.
+            If not provided, the model will be inferred from the function's type hints.
+            The primary reason why you might want to provide a model in
+            Note that the name and docstring of the model takes precedence over the
+            name and docstring of the function.
+
+        Returns
+        -------
+        Tool
+            A new Tool instance wrapping the provided function.
+
+        Raises
+        ------
+        ValueError
+            If there is a mismatch between model fields and function parameters.
+        """
+
+        if model is None:
+            model = func_to_basemodel(func)
+
+        # Throw if there is a mismatch between the model and the function parameters
+        params = inspect.signature(func).parameters
+        fields = model.model_fields
+        diff = set(params) ^ set(fields)
+        if diff:
+            raise ValueError(
+                f"`model` fields must match tool function parameters exactly. "
+                f"Fields found in one but not the other: {diff}"
+            )
+
+        params = basemodel_to_param_schema(model)
+
+        return cls(
+            func=func,
+            name=model.__name__ or func.__name__,
+            description=model.__doc__ or func.__doc__ or "",
+            parameters=params,
+        )
+
+    @classmethod
+    def from_mcp(
+        cls: type["Tool"],
+        session: "MCPClientSession",
+        mcp_tool: "MCPTool",
+    ) -> "Tool":
+        """
+        Create a Tool from an MCP tool
+
+        Parameters
+        ----------
+        session
+            The MCP client session to use for calling the tool.
+        mcp_tool
+            The MCP tool to wrap.
+
+        Returns
+        -------
+        Tool
+            A new Tool instance wrapping the MCP tool.
+        """
+
+        async def _call(**args: Any) -> AsyncGenerator[ContentToolResult, None]:
+            result = await session.call_tool(mcp_tool.name, args)
+
+            # Raise an error if the tool call resulted in an error. It doesn't seem to be
+            # very well defined how to get at the error message, but it appears that it gets
+            # stored in the `text` attribute of the content. Also, empirically, the error
+            # message seems to include `Error executing tool {tool_name}: ...`, so
+            if result.isError:
+                err_msg = getattr(
+                    result.content[0],
+                    "text",
+                    f"Error executing tool {mcp_tool.name}.",
+                )
+                raise RuntimeError(err_msg)
+
+            for content in result.content:
+                if content.type == "text":
+                    yield ContentToolResult(value=content.text)
+                elif content.type == "image":
+                    if content.mimeType not in (
+                        "image/png",
+                        "image/jpeg",
+                        "image/webp",
+                        "image/gif",
+                    ):
+                        raise ValueError(
+                            f"Unsupported image MIME type: {content.mimeType}"
+                        )
+
+                    yield ContentToolResultImage(
+                        value=content.data,
+                        mime_type=content.mimeType,
+                    )
+                elif content.type == "resource":
+                    from mcp.types import TextResourceContents
+
+                    resource = content.resource
+                    if isinstance(resource, TextResourceContents):
+                        blob = resource.text.encode("utf-8")
+                    else:
+                        blob = resource.blob.encode("utf-8")
+
+                    yield ContentToolResultResource(
+                        value=blob, mime_type=content.resource.mimeType
+                    )
+                else:
+                    raise RuntimeError(f"Unexpected content type: {content.type}")
+
+        params = mcp_tool_input_schema_to_param_schema(mcp_tool.inputSchema)
+
+        return cls(
+            func=_utils.wrap_async(_call),
+            name=mcp_tool.name,
+            description=mcp_tool.description or "",
+            parameters=params,
+        )
 
 
 class ToolRejectError(Exception):
@@ -160,14 +310,6 @@ def func_to_basemodel(func: Callable) -> type[BaseModel]:
 
 
 def basemodel_to_param_schema(model: type[BaseModel]) -> dict[str, object]:
-    try:
-        import openai
-    except ImportError:
-        raise ImportError(
-            "The openai package is required for this functionality. "
-            "Please install it with `pip install openai`."
-        )
-
     # Lean on openai's ability to translate BaseModel.model_json_schema()
     # to a valid tool schema (this wouldn't be impossible to do ourselves,
     # but it's fair amount of logic to substitute `$refs`, etc.)
@@ -177,10 +319,27 @@ def basemodel_to_param_schema(model: type[BaseModel]) -> dict[str, object]:
     if "parameters" not in fn:
         raise ValueError("Expected `parameters` in function definition.")
 
-    params = fn["parameters"]
+    params = rm_param_titles(fn["parameters"])
 
-    # For some reason, openai (or pydantic?) wants to include a title
-    # at the model and field level. I don't think we actually need or want this.
+    return params
+
+
+def mcp_tool_input_schema_to_param_schema(
+    input_schema: dict[str, Any],
+) -> dict[str, object]:
+    params = rm_param_titles(input_schema)
+
+    if "additionalProperties" not in params:
+        params["additionalProperties"] = False
+
+    return params
+
+
+def rm_param_titles(
+    params: dict[str, object],
+) -> dict[str, object]:
+    # For some reason, pydantic wants to include a title at the model and field
+    # level. I don't think we actually need or want this.
     if "title" in params:
         del params["title"]